pyproject.toml
[tool.poetry]
name = "owlgram"
version = "0.1.0"
description = "Owlgram API"
authors = ["D0Nater <denaddressforwork@gmail.com>"]
readme = "README.md"

[tool.poetry.scripts]
owlgram = "owlgram.__main__:main"

[tool.poetry.dependencies]
python = "^3.12"


# FastAPI
fastapi = "^0.111.0"                                   # Web-framework
pydantic = "^2.7"                                      # Serialization-framework
python-multipart = "^0.0.9"                            # Multipart form files handling
pyfa-converter-v2 = "^2.0.0rc1"                        # Utility to convert Pydantic models to FastAPI dependencies
uvicorn = { extras = ["standard"], version = "^0.30" } # ASGI web-server

# Database
sqlalchemy = "^2.0"      # Database-framework
asyncpg = "^0.29"        # Async postgres driver
alembic = "^1.13"        # Database migrations
python-dateutil = "^2.9" # Date handling
pytz = "^2024.1"         # Timezone handling

# Caching
redis = { version = "^5.0", extras = ["hiredis"] } # Redis

# CLI
click = "^8.1" # CLI framework

# Other
pydantic-settings = "^2.3.3" # Settings framework
pyjwt = "^2.8.0"             # JWT


[tool.poetry.group.dev.dependencies]
types-click = "^7.1.8" # For Click mypy support
black = "^24.3.0" # Code formatter
isort = "^5.12.0" # Import sorting
flake8 = "^7.0.0" # Linter
flake8-docstrings = "^1.6.0" # Docstring linter
mypy = "^1.10.0" # Static type checking
pre-commit = "^3.7.1" # Git pre commit hooks


[tool.black]
line-length = 120
target-version = ['py312']

[tool.isort]
line_length = 120
atomic = true
include_trailing_comma = true
use_parentheses = true
profile = "black"
lines_before_imports = 1
lines_after_imports = 2
combine_as_imports = true
multi_line_output = 3


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
